% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{appendSEDOLCheckDigit}
\alias{appendSEDOLCheckDigit}
\title{SEDOL check digits}
\usage{
appendSEDOLCheckDigit(x)
}
\arguments{
\item{x}{Input containing SEDOLs (character)}
}
\value{
Character.
}
\description{
Append a check-digit to input SEDOLs
}
\details{
Implemented the algorithm for computing the SEDOL check digit, as
described on Wikipedia.
}
\examples{
sedols <- c("208942", "298356", "BFSSDS", "270502")
appendSEDOLCheckDigit(sedols)
}
\references{
Wiki page: \url{https://en.wikipedia.org/wiki/SEDOL}
}
