% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cmp}
\alias{cmp}
\title{Quickie Visual Comparison}
\usage{
cmp(a, b, xlab = "A", ylab = "B", ...)
}
\arguments{
\item{a}{First array}

\item{b}{Second array}

\item{xlab}{Default x-axis label}

\item{ylab}{Default y-axis label}

\item{...}{Additional arguments passed to \code{plot}}
}
\description{
Quick X-Y plot to compare contents of two vectors. If vectors have row
names, use that to key them for comparison.
}
\details{
I constantly find myself comparing two vectors to see if they are equal.
A quick X-Y plot is a great way to do this visually and is better than
correlation at spotting outliers or other structure.

Usually pass two column-like things (matrices or data frames), usually
keyed by row names. But can pass two list-like things (e.g. numerics) or
even a two-column data frame/matrix. (Columns to the right of the first
two are ignored).
}
\examples{
x <- rnorm(50)
y <- 0.3*x^2
cmp(x, y)

x <- data.frame(x=rnorm(26), y=rnorm(26), row.names=toupper(letters))
cmp(x)

x <- data.frame(x=rnorm(26), row.names=toupper(letters))
y <- data.frame(y=rnorm(26), row.names=toupper(letters))
cmp(x, y)
}
