% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateUtil.R
\name{addDays}
\alias{addDays}
\alias{subtractDays}
\alias{addBizDays}
\alias{subtractBizDays}
\alias{addMonths}
\alias{subtractMonths}
\title{Date arithmetic}
\usage{
addDays(date, n = 1)

subtractDays(date, n = 1)

addBizDays(date, n = 1, hol = NULL)

subtractBizDays(date, n = 1, hol = NULL)

addMonths(date, n = 1)

subtractMonths(date, n = 1)
}
\arguments{
\item{date}{A 'datelike' object}

\item{n}{An integer number of days/months to add/subtract (default 1)}

\item{hol}{A 'datelike' object containing holiday dates}
}
\description{
Bunch of convenience functions to do simple arithmetic on dates.
}
\details{
Increment or decrement the vector 'date' forward or backward by
'n' day(s), 'n' business day(s), or 'n' months. Business days
are considered to be Monday through Friday. A vector of holiday
dates can be passed in.
}
\examples{
\dontrun{
x <- seq(from=as.Date("2001-01-01"), to=as.Date("2001-01-31"), by="day")
ax <- addDays(x, 3)
difftime(x, ax)

x <- seq(from=as.Date("2001-01-01"), to=as.Date("2001-01-31"), by="day")

sx <- subtractDays(x, 3)
difftime(x, sx)

x <- seq(from=as.Date("2001-01-01"), to=as.Date("2002-06-01"), by="months")
h <- as.Date("2001-07-02")
ax <- addBizDays(x, hol=h)
}
}
\seealso{
is.DateLike
}
