% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateUtil.R
\name{Date Manipulation}
\alias{Date Manipulation}
\alias{firstDayInMonth}
\alias{lastDayInMonth}
\alias{firstBizDayInMonth}
\alias{lastBizDayInMonth}
\alias{previousMonthEnd}
\title{Date Manipulation}
\usage{
firstDayInMonth(x)

lastDayInMonth(x)

firstBizDayInMonth(x, hol = NULL)

lastBizDayInMonth(x, hol = NULL)

previousMonthEnd(x)
}
\arguments{
\item{x}{A 'datelike' object}

\item{hol}{A 'datelike' object containing holiday dates}
}
\description{
Move date input to convenient dates at the beginning or end
of the the month.
}
\details{
Business days
are considered to be Monday through Friday. A vector of holiday
dates can be passed in.
}
\seealso{
is.DateLike
}
