% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{expWts}
\alias{expWts}
\title{Exponential weights}
\usage{
expWts(length = 10, halfLife = 6, max = 1)
}
\arguments{
\item{length}{Length of output array (default=10)}

\item{halfLife}{Half-life, in periods (default=6)}

\item{max}{Maximum weight (default=1)}
}
\value{
Numeric
}
\description{
Compute exponential weights for a specified half-life.
}
\details{
Generate array of exponential weights of length \code{length},
and half-life \code{halfLife}. Maximum weight is \code{max} (default 1.0).
}
\examples{
expWts()
expWts(50, 30)
}
