% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateUtil.R
\name{is.DateLike}
\alias{is.DateLike}
\title{Check that input is a date}
\usage{
is.DateLike(x)
}
\arguments{
\item{x}{Input to test}
}
\value{
Boolean
}
\description{
Test if input is a date-like object
}
\details{
Test if input is one of my favorite date-like objects. In other words,
a \code{Date}, or a \code{POSIX*}.
}
\examples{
dtstr <- c("1969-08-03", "2011-06-24")

is.DateLike("AAA")
is.DateLike(251)
is.DateLike(as.Date(dtstr))
is.DateLike(strptime(dtstr, "\\\%Y-\\\%m-\\\%d"))
}
