% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateUtil.R
\name{isBizday}
\alias{isBizday}
\title{Check for business days}
\usage{
isBizday(x, hol = NULL, wday = 1:5)
}
\arguments{
\item{x}{A 'datelike' object}

\item{hol}{A 'datelike' object containing holiday dates}

\item{wday}{Weekdays to use (integer; default: Monday - Friday)}
}
\value{
Boolean
}
\description{
Which of input 'x' is a "business" day? Standard def'n of weekdays is
1:5 = Monday - Friday. Returns logical.
}
