% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ll}
\alias{ll}
\title{List Objects with Detail}
\usage{
ll(pattern, pos = 1, order.by = "size", decreasing = TRUE)
}
\arguments{
\item{pattern}{Optional \code{\link{regular expression}}. Only
names matching \code{pattern} are returned.}

\item{pos}{Specify environment as a position in the search list.}

\item{order.by}{Order by \sQuote{type} or \sQuote{size}}

\item{decreasing}{List in decreasing/increasing order of /code{order.by}}
}
\description{
List objects in w/ more detail than ls(). Similar to a unix-like
\code{ls -l}.
}
\details{
List objects in w/ more detail than ls(). Modified from a
\href{http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}{Stackoverflow question}
}
\examples{
require(graphics)
ll()
}
