% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{matrixFromDatedFiles}
\alias{matrixFromDatedFiles}
\title{Matrix from dated files}
\usage{
matrixFromDatedFiles(fls, dir, filePat = "*", col = 1, ...)
}
\arguments{
\item{fls}{Files (character) to process}

\item{dir}{Directory (character)}

\item{filePat}{File pattern (regexpr)}

\item{col}{Column description (integer or character)}

\item{...}{Additional arguments passed to \code{read.table}}
}
\value{
A matrix
}
\description{
Create a single matrix from data in separate (dated) files.
}
\details{
This is a general-purpose function to parse dated files to a data matrix.
There's nothing here about the file-naming convention; that's the
responsibility of the caller. (Columns get labelled exactly like the
input files). Missing data padded w/ 'NA'.
Load data from columns in the dated files. Can either specify files
directly or specify a directory 'dir' and file-pattern 'filePat' (regexpr).
Column description 'col', which can be either an integer column index
or a character which will be passed to grep to match a column header.
Arguments like 'skip', 'header', 'sep' & 'stringsAsFactors'
are passed to 'read.table'.

This is a general-purpose function to parse dated files to a data matrix.
There's nothing here about the file-naming convention; that's the
responsibility of the caller. (Columns get labelled exactly like the
input files).
}
\examples{
\dontrun{
dat <- matrixFromDatedFiles('data', 'dat*', sep=",", header=T)
}
}
