% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcLib.R
\docType{package}
\name{pcLib-package}
\alias{pcLib}
\alias{pcLib-package}
\title{pcLib package}
\description{
Pete's miscellaneous convenience functions.
}
\details{
Variety of utility funcs I use regularly.
}
\author{
P. Christiansen, \email{pchristi@alumni.stanford.edu}
}
