% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsUtil.R
\name{plotz}
\alias{plotz}
\title{Plot a \code{zoo}}
\usage{
plotz(x, ...)
}
\arguments{
\item{x}{Input \code{zoo}}

\item{...}{Additional parameters passed to \code{plot}}
}
\description{
Plot an input \code{zoo}, embedding my personal preferences
}
