% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{resources}
\alias{resources}
\title{Report Resource Usage}
\usage{
resources(expr)
}
\arguments{
\item{expr}{An object to be evaluated (e.g. function call).}
}
\description{
Report on memory usage (not Mac OS X yet) and processing time.
}
\details{
Evaluate \code{expr} and print processing time, session time (both in
seconds) and memory usage (certain platforms only).
Gives CPU & memory resources used in evaluating 'expr'
(modified from Venebles & Ripley).
}
\examples{
A <- matrix(rnorm(1e6), 1e3)
resources( svd(A) )
}
