% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsUtil.R
\name{sampleZooMonthly}
\alias{sampleZooMonthly}
\title{Sample a \code{zoo}}
\usage{
sampleZooMonthly(x, register = c("first", "last"))
}
\arguments{
\item{x}{An input \code{zoo} object}

\item{register}{Flag to indicate where to register the output:
Either 'first' or 'last'}
}
\value{
A \code{zoo}
}
\description{
Sample an input \code{zoo} onto a monthly frequency.
}
\details{
Sub-sample a \code{zoo} from (usually) daily to a monthly frequency. Register
the output on either first or last day of the month. This simply extracts
the records from the input that are closest to either the beginning or ending
day of each month covered. It's a "coarsening" of a \code{zoo} with a
frequency higher than monthly (e.g. daily to monthly).
}
\examples{
library(zoo)

dts1 <- seq(as.Date("2012-08-01"), as.Date("2013-09-30"), "days")
dts2 <- seq(strptime("20120801", "\\\%Y\\\%m\\\%d"),
strptime("20130930", "\\\%Y\\\%m\\\%d"), "day")

nr <- length(dts1)
z1 <- zoo(matrix(rnorm(nr*4), nr, 4), dts1)
z2 <- zoo(matrix(rnorm(nr*4), nr, 4), dts2)

sampleZooMonthly(z1)
sampleZooMonthly(z2, "last")
}
