% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsUtil.R
\name{scoreZoos}
\alias{scoreZoos}
\title{Time-series score a \code{zoo}}
\usage{
scoreZoos(x, hl, na.rm = TRUE)
}
\arguments{
\item{x}{Input \code{zoo}}

\item{hl}{Half life (in whatever frequeny indexes \code{x})}

\item{na.rm}{Remove missing values? (Boolean)}
}
\value{
A \code{zoo} with the scored output
}
\description{
Time-series score \code{x} using an expanding window & weighting stats
estimates using exponentially decaying weights with half-life
\code{hl}.
}
\details{
\code{scoreZoos} time-series scores the input zoo using an expanding
window and exponentially weighting the estimated mean and standard deviation
using the supplied half-life. Units of the half-life are however the input
series is indexed. If no \code{hl} argument is supplied, uses
equally-weighted estimates.
}
