% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize}
\usage{
winsorize(x, mult = 3)
}
\arguments{
\item{x}{a numeric matrix-like object}

\item{mult}{multiplier}
}
\value{
Numeric.
}
\description{
"Windorize" the (numeric) input by truncating outliers
}
\details{
Handle outliers in the columns of a numeric matrix by setting them
equal to \code{mult} times the distribution median.
Simple handling of data outliers. Less involved than iterative
re-computation of sd & truncation.

Inspired by blog post here: \url{https://www.r-bloggers.com/winsorization/}.
}
\examples{
norm1 <- rnorm(100)
cauch1 <- rcauchy(100)
dat <- cbind(norm1, cauch1)

datw <- winsorize(dat)
}
